requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.25.2
scipy==1.11.2
matplotlib==3.7.2
Pillow==10.0.0
